backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
pinecone==6.0.1
sentence-transformers==2.2.2
openai>=1.6.1,<2.0.0
pypdf==3.17.0
docx2txt==0.8
langchain==0.0.335
langchain-openai==0.0.2
numpy==1.26.2
scikit-learn==1.3.2
pillow==10.1.0 